% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tPAUC.R
\name{data_crossingdc}
\alias{data_crossingdc}
\title{Title}
\usage{
data_crossingdc(n, vt, vc, vm, vxs, u)
}
\arguments{
\item{n}{}

\item{vt}{}

\item{vc}{}

\item{vm}{}

\item{vxs}{}

\item{u}{}
}
\value{
reshaped data for easy optimization
}
\description{
Title
}
