% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc.R
\name{proc}
\alias{proc}
\title{Partial AUC Estimation and Inference}
\usage{
proc(response, predictor, threshold = 0.9, method = "MW", ci = TRUE,
  cp = 0.95, plot = TRUE, smooth = FALSE)
}
\arguments{
\item{response}{a factor, numeric or character vector of responses; 
typically encoded with 0 (negative) and 1 (positive). 
Only two classes can be used in a ROC curve. If its levels are not 0/1,
the first level will be defaultly regarded as negative.}

\item{predictor}{a numeric vector of the same length than response, containing the predicted value of each observation. An ordered factor is coerced to a numeric.}

\item{threshold}{numeric; false positive rate (FPR) constraint.}

\item{method}{methods to estimate FPR-pAUC. \code{MW}: Mann-Whitney statistic. \code{expect}: method in (2.2) \href{http://www.ncbi.nlm.nih.gov/pubmed/20729218}{Wang and Chang, 2011}. \code{jackknife}: jackknife method in \href{http://www3.stat.sinica.edu.tw/ss_newpaper/SS-13-367_na.pdf}{Yang et al., 2016}.}

\item{ci}{logic; compute the confidence interval of estimation?}

\item{cp}{numeric; coverage probability of confidence interval.}

\item{plot}{logic; plot the ROC curve?}

\item{smooth}{if \code{TRUE}, the ROC curve is passed to \code{\link[pROC]{smooth}} to be smoothed.}
}
\value{
Estimation and Inference of FPR partial AUC.
}
\description{
Estimate and infer the area of region under ROC curve with pre-specific FPR constraint (FPR-pAUC). See \href{http://www3.stat.sinica.edu.tw/ss_newpaper/SS-13-367_na.pdf}{Yang et al., 2016} for details.
}
\details{
This function estimates and infers FPR partial AUC given response, predictor and pre-specific FPR constraint. The plot of corresponding ROC curve with pre-specific FPR is generated.
         \code{MW}: Mann-Whitney statistic. \code{expect}: method in (2.2) \href{http://www.ncbi.nlm.nih.gov/pubmed/20729218}{Wang and Chang, 2011}. \code{jackknife}: jackknife method in \href{http://www3.stat.sinica.edu.tw/ss_newpaper/SS-13-367_na.pdf}{Yang et al., 2016}.
}
\examples{

library('pROC')
data(aSAH)
proc(aSAH$outcome, aSAH$s100b,threshold=0.9, method='expect',ci=TRUE, cp=0.95, plot=TRUE)

}
\author{
Hanfang Yang, Kun Lu, Xiang Lyu, Feifang Hu.
}
\seealso{
\code{\link[pROC]{roc}}, \code{\link[tpAUC]{tproc.est}}, \code{\link[tpAUC]{proc.est}}, \code{\link[tpAUC]{proc.ci}}
}

