\name{tPAUC-package}
\alias{tPAUC-package}
\alias{tPAUC}
\docType{package}
\title{
\packageTitle{tPAUC}
}
\description{
\packageDescription{tPAUC}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{tPAUC}
\packageIndices{tPAUC}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{tPAUC}

Maintainer: \packageMaintainer{tPAUC}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS ~~
~~ in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
