% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpAUC-package.r
\docType{package}
\name{tpAUC}
\alias{tpAUC}
\alias{tpAUC-package}
\title{Estimation and Inference of Two-Way Partial AUC, FPR partial AUC and FNR partial ODC}
\description{
Tools of estimation and inference of two-way partial AUC, FPR partial AUC and FNR partial ODC. Methods are proposed in \href{http://arxiv.org/abs/1508.00298}{Yang et al., 2016} and \href{http://www3.stat.sinica.edu.tw/ss_newpaper/SS-13-367_na.pdf}{Yang et al., 2016}, including \code{Mann-Whitney Statistic}, \code{jackknife method}, etc.
}
\details{
\tabular{ll}{
  Package: \tab tpAUC \cr
  Type: \tab Package \cr
  Date \tab 2016-08-20 \cr
  License: \tab GPL (>= 2) \cr
}
}
\author{
\tabular{l}{
  Hanfang Yang, Kun Lu, Xiang Lyu, Feifang Hu. \cr
  Maintainer: Xiang Lyu  <lyu17@purdue.edu> 
}
}
\references{
\tabular{l}{
       Wang Z, Chang Y. Marker selection via maximizing the partial area under the ROC curve of linear risk scores[J]. \emph{Biostatistics}, 2011, 12(2): 369-385. \cr
       Yang H, Lu K, Lyu X, Hu F. Two-Way Partial AUC and Its Properties[J]. \emph{arXiv:1508.00298}, 2016.   \cr
       Yang H, Lu K, Zhao Y. A nonparametric approach for partial areas under ROC curves and ordinal dominance curves. \emph{Statistica Sinica}. 2016.
 }
}

